refactor translation so we don't have double loop
