+++ src/main/routes/features/dashboard/dashboardController.ts
@@ -5,13 +5,15 @@ import {DASHBOARD_URL} from '../../urls';
 import {AppRequest, UserDetails} from 'models/AppRequest';
 import {getOcmcDraftClaims} from '../../../app/client/legacyDraftStoreClient';
 import {DashboardClaimantItem, DashboardDefendantItem} from '../../../common/models/dashboard/dashboardItem';
-import {getClaimsForClaimant, getClaimsForDefendant} from 'services/features/dashboard/dashboardService';
+import {CivilServiceClient} from 'client/civilServiceClient';
 
+const civilServiceApiBaseUrl = config.get<string>('services.civilService.url');
+const civilServiceClient: CivilServiceClient = new CivilServiceClient(civilServiceApiBaseUrl);
 const ocmcBaseUrl = config.get<string>('services.cmc.url');
 
 function renderPage(res: Response, claimsAsClaimant: DashboardClaimantItem[], claimDraftSaved: DashboardClaimantItem,
   claimsAsDefendant: DashboardDefendantItem[], responseDraftSaved: boolean,
-  paginationArgumentClaimant: object, paginationArgumentDefendant: object): void {
+  paginationArgumentClaimant: object, paginationArgumentDefendant: object, lang: string | unknown): void {
   res.render('features/dashboard/dashboard', {
     claimsAsClaimant,
     claimDraftSaved,
@@ -19,6 +21,7 @@ function renderPage(res: Response, claimsAsClaimant: DashboardClaimantItem[], cl
     responseDraftSaved,
     paginationArgumentClaimant,
     paginationArgumentDefendant,
+    lang,
     newOcmcClaimUrl: `${ocmcBaseUrl}/eligibility`,
   });
 }
@@ -26,15 +29,16 @@ function renderPage(res: Response, claimsAsClaimant: DashboardClaimantItem[], cl
 const dashboardController = Router();
 
 dashboardController.get(DASHBOARD_URL, async function (req, res) {
+  const lang = req.query.lang ? req.query.lang : req.cookies.lang;
   const appRequest = <AppRequest> req;
   const user: UserDetails = appRequest.session.user;
-  const claimsAsClaimant : DashboardClaimantItem[] = await getClaimsForClaimant(req);
-  const claimsAsDefendant : DashboardDefendantItem[] = await getClaimsForDefendant(req);
+  const claimsAsClaimant : DashboardClaimantItem[] = await civilServiceClient.getClaimsForClaimant(appRequest);
+  const claimsAsDefendant : DashboardDefendantItem[] = await civilServiceClient.getClaimsForDefendant(appRequest);
   const claimDraftSaved = await getOcmcDraftClaims(user?.accessToken);
   const responseDraftSaved = false;
   const paginationArgumentClaimant: object = {};
   const paginationArgumentDefendant: object = {};
-  renderPage(res, claimsAsClaimant, claimDraftSaved, claimsAsDefendant, responseDraftSaved, paginationArgumentClaimant, paginationArgumentDefendant);
+  renderPage(res, claimsAsClaimant, claimDraftSaved, claimsAsDefendant, responseDraftSaved, paginationArgumentClaimant, paginationArgumentDefendant, lang);
 });
 
 export default dashboardController;
