+++ src/main/views/features/dashboard/dashboard.njk
@@ -104,7 +104,7 @@
                 </td>
                 <td class="govuk-table__cell">{{ item.claimantName }}</td>
                 <td class="govuk-table__cell">£{{ item.claimAmount }}</td>
-                <td class="govuk-table__cell">{{ item.translatedStatus }} </td>
+                <td class="govuk-table__cell">{{ item.getStatus(lang) }} </td>
               </tr>
             {% endfor %}
           </tbody>
